feat(Select): Refactor Select into more split up components
