+++ packages/components/src/components/Select/style.tsx
@@ -66,28 +66,6 @@ const StyledWrapper = styled.div<WrapperPropsType>`
     outline: none;
     display: inline-block;
     position: relative;
-    z-index: 1;
-
-    &::before {
-        content: '';
-        z-index: -1;
-        border-radius: ${({ theme }) => theme.Select.common.borderRadius};
-        ${({ open }) =>
-            open
-                ? `
-                border-bottom-left-radius: 0;
-                border-bottom-right-radius: 0;
-                `
-                : ''}
-        border: solid ${({ theme }) => theme.Select.common.borderColor};
-        border-width: ${({ open }) => (open ? `1px 1px 0` : `0`)};
-        background: ${({ theme }) => theme.Select.searchWrapper.background};
-        position: absolute;
-        top: ${({ open }) => (open ? `-${INNER_OFFSET}px` : '0')};
-        left: ${({ open }) => (open ? `-${INNER_OFFSET}px` : '0')};
-        right: ${({ open }) => (open ? `-${INNER_OFFSET}px` : '0')};
-        bottom: ${({ open }) => (open ? `-${INNER_OFFSET}px` : '0')};
-    }
 `;
 
 const StyledCaret = styled.div`
@@ -110,6 +88,7 @@ const StyledSelection = styled.div<SelectionProps>`
     transition: color 150ms;
     -webkit-font-smoothing: antialiased;
     -moz-osx-font-smoothing: grayscale;
+    user-select: none;
 `;
 
 const composeSelectTheme = (themeTools: ThemeTools): SelectThemeType => {
